shopfloor: make auto_post menu option available to all scenarios
