+++ shopfloor_reception/__manifest__.py
@@ -12,7 +12,6 @@
     "depends": ["shopfloor"],
     "data": [
         "data/shopfloor_scenario_data.xml",
-        "views/shopfloor_menu.xml",
     ],
     "demo": [
         "demo/stock_picking_type_demo.xml",
