--- shopfloor_reception/models/__init__.py
@@ -1,1 +1,0 @@
-from . import shopfloor_menu
