--- shopfloor_reception/views/shopfloor_menu.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<odoo>
-    <record id="shopfloor_menu_form_view" model="ir.ui.view">
-        <field name="model">shopfloor.menu</field>
-        <field name="inherit_id" ref="shopfloor.shopfloor_menu_form_view" />
-        <field name="arch" type="xml">
-            <group name="options" position="inside">
-              <group
-                    name="auto_post_line"
-                    attrs="{'invisible': [('auto_post_line_is_possible', '=', False)]}"
-                >
-                  <field name="auto_post_line_is_possible" invisible="1" />
-                  <field name="auto_post_line" />
-              </group>
-            </group>
-        </field>
-    </record>
-</odoo>
